+++ api/tests/server.test.ts
@@ -1,15 +1,12 @@
 import type { MockInstance } from "vitest";
 import { describe, it, beforeEach, expect, vi } from "vitest";
 import { AutoRouter } from "itty-router";
-import { InteractionType } from "discord-api-types/v10";
 import type * as HaloInfiniteApi from "halo-infinite-api";
 import { HaloInfiniteClient, StaticXstsTicketTokenSpartanTokenProvider } from "halo-infinite-api";
 import { installFakeServicesWith } from "../services/fakes/services";
 import { Server } from "../server";
-import { getCommands } from "../commands/commands";
 import { aFakeEnvWith } from "../base/fakes/env.fake";
 import { aFakeHaloInfiniteClient } from "../services/halo/fakes/infinite-client.fake";
-import { pingInteraction } from "../services/discord/fakes/data";
 import type { TokenInfo } from "../services/xbox/types";
 
 vi.mock("halo-infinite-api", async (importOriginal) => {
@@ -33,7 +30,6 @@ describe("Server", () => {
     server = new Server({
       router: AutoRouter(),
       installServices,
-      getCommands,
     });
   });
 
@@ -118,7 +114,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
 
       const req = new Request("http://localhost/proxy/halo-infinite", {
@@ -164,7 +159,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
       const req = new Request("http://localhost/proxy/halo-infinite", {
         method: "POST",
@@ -216,7 +210,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
 
       const req = new Request("http://localhost/proxy/halo-infinite", {
@@ -250,7 +243,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
 
       const req = new Request("http://localhost/proxy/halo-infinite", {
@@ -297,7 +289,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
       const req = new Request("http://localhost/proxy/halo-infinite", {
         method: "POST",
@@ -415,7 +406,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
       const req = new Request("http://localhost/proxy/halo-infinite", {
         method: "POST",
@@ -448,161 +438,6 @@ describe("Server", () => {
     });
   });
 
-  describe("POST /interactions", () => {
-    it("returns 401 when Discord verification fails", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        const mockVerifyDiscordRequest = vi.spyOn(services.discordService, "verifyDiscordRequest").mockResolvedValue({
-          isValid: false,
-          rawBody: "invalid-body",
-        });
-        return { ...services, verifyDiscordRequest: mockVerifyDiscordRequest };
-      });
-
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-
-      const req = new Request("http://localhost/interactions", {
-        method: "POST",
-        body: JSON.stringify({ type: 1 }),
-        headers: {
-          "content-type": "application/json",
-          "x-signature-ed25519": "invalid-signature",
-          "x-signature-timestamp": "123456789",
-        },
-      });
-
-      const ctx: Partial<EventContext<Env, "", unknown>> = {
-        waitUntil: vi.fn(),
-        passThroughOnException: vi.fn(),
-      };
-
-      const res = (await server.router.fetch(req, env, ctx as EventContext<Env, "", unknown>)) as Response;
-      expect(res.status).toBe(401);
-      const text = await res.text();
-      expect(text).toBe("Bad request signature.");
-    });
-
-    it("handles PING interaction successfully", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.discordService, "verifyDiscordRequest").mockResolvedValue({
-          isValid: true,
-          interaction: pingInteraction,
-          rawBody: JSON.stringify(pingInteraction),
-        });
-        return services;
-      });
-
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-
-      const req = new Request("http://localhost/interactions", {
-        method: "POST",
-        body: JSON.stringify(pingInteraction),
-        headers: {
-          "content-type": "application/json",
-          "x-signature-ed25519": "valid-signature",
-          "x-signature-timestamp": "123456789",
-        },
-      });
-
-      const ctx: Partial<EventContext<Env, "", unknown>> = {
-        waitUntil: vi.fn(),
-        passThroughOnException: vi.fn(),
-      };
-
-      const res = (await server.router.fetch(req, env, ctx as EventContext<Env, "", unknown>)) as Response;
-      expect(res.status).toBe(200);
-      const body = await res.json();
-      expect(body).toHaveProperty("type", InteractionType.Ping);
-    });
-
-    it("returns 500 on internal error", async () => {
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.discordService, "verifyDiscordRequest").mockRejectedValue(new Error("Internal failure"));
-        return services;
-      });
-
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-
-      const req = new Request("http://localhost/interactions", {
-        method: "POST",
-        body: JSON.stringify({ type: 1 }),
-        headers: {
-          "content-type": "application/json",
-          "x-signature-ed25519": "valid-signature",
-          "x-signature-timestamp": "123456789",
-        },
-      });
-
-      const ctx: Partial<EventContext<Env, "", unknown>> = {
-        waitUntil: vi.fn(),
-        passThroughOnException: vi.fn(),
-      };
-
-      const res = (await server.router.fetch(req, env, ctx as EventContext<Env, "", unknown>)) as Response;
-      expect(res.status).toBe(500);
-      const text = await res.text();
-      expect(text).toBe("Internal error");
-    });
-
-    it("calls waitUntil when jobToComplete is provided", async () => {
-      const jobToCompleteMock = vi.fn(async () => {
-        // Empty implementation
-      });
-      const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
-        const services = installFakeServicesWith({ env });
-        vi.spyOn(services.discordService, "verifyDiscordRequest").mockResolvedValue({
-          isValid: true,
-          interaction: pingInteraction,
-          rawBody: JSON.stringify(pingInteraction),
-        });
-        vi.spyOn(services.discordService, "handleInteraction").mockReturnValue({
-          response: new Response(JSON.stringify({ type: 1 })),
-          jobToComplete: jobToCompleteMock,
-        });
-        return services;
-      });
-
-      server = new Server({
-        router: AutoRouter(),
-        installServices: localInstallServices,
-        getCommands,
-      });
-
-      const req = new Request("http://localhost/interactions", {
-        method: "POST",
-        body: JSON.stringify(pingInteraction),
-        headers: {
-          "content-type": "application/json",
-          "x-signature-ed25519": "valid-signature",
-          "x-signature-timestamp": "123456789",
-        },
-      });
-
-      const waitUntilSpy = vi.fn();
-      const ctx: Partial<EventContext<Env, "", unknown>> = {
-        waitUntil: waitUntilSpy,
-        passThroughOnException: vi.fn(),
-      };
-
-      await server.router.fetch(req, env, ctx as EventContext<Env, "", unknown>);
-      expect(waitUntilSpy).toHaveBeenCalledWith(expect.any(Promise));
-    });
-  });
-
   describe("POST /neatqueue", () => {
     it("returns 401 when verification fails", async () => {
       const localInstallServices = vi.fn<typeof installFakeServicesWith>(() => {
@@ -617,7 +452,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
 
       const req = new Request("http://localhost/neatqueue", {
@@ -650,7 +484,6 @@ describe("Server", () => {
       server = new Server({
         router: AutoRouter(),
         installServices: localInstallServices,
-        getCommands,
       });
 
       const req = new Request("http://localhost/neatqueue", {
